+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 ### Bug fixes
 
 ### Deprecations
+- Use `XCUITest` automation name instead of non or `appium` automation name for iOS platform
+    - Since Appium no longer supports under Xcode 8
 
 ## [2.0.1] - 2018-09-01
 ### Enhancements